aws-backend/sm-ep-embed-image/code/requirements.txt
transformers[torch]
pillow
sagemaker-inference
